requirements.txt
fastapi
uvicorn
supabase
google-generativeai
httpx
python-dotenv
